+++ bit_manip/bit_manip.hpp
@@ -1,12 +1,13 @@
 #ifndef BIT_MANIP
 #define BIT_MANIP
 
+#include <cstdint>
 #include <string>
 
 using namespace std;
 
-string binary(int n);
-int count_1bit(int n);
-int degree(int n);
+string binary(uint64_t n);
+int count_1bit(uint64_t n);
+int degree(uint64_t n);
 
 #endif // !BIT_MANIP
